# perf: improve idgenerator (about 25-30 % faster)