--- a/README.md
+++ b/README.md
@@ -89,7 +89,7 @@ Setting this to `false` can save storage space and comply with the EU cookie law
 
 ##### idGenerator(request) (optional)
 
-Function used to generate new session IDs..
+Function used to generate new session IDs.
 Custom implementation example:
 ```js
 const uid = require('uid-safe').sync