--- a/lib/fastifySession.js
+++ b/lib/fastifySession.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const fp = require('fastify-plugin')
-const uid = require('uid-safe').sync
+const idGenerator = require('./idGenerator')
 const Store = require('./store')
 const Session = require('./session')
 const cookieSignature = require('cookie-signature')
@@ -180,10 +180,6 @@ function checkOptions (options) {
   if (Array.isArray(options.secret) && options.secret.length === 0) {
     return new Error('at least one secret is required')
   }
-}
-
-function idGenerator () {
-  return uid(24)
 }
 
 function ensureDefaults (options) {