--- a/lib/fastifySession.js
+++ b/lib/fastifySession.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const fp = require('fastify-plugin')
-const idGenerator = require('./idGenerator')
+const idGenerator = require('./idGenerator')()
 const Store = require('./store')
 const Session = require('./session')
 const cookieSignature = require('cookie-signature')