--- a/lib/idGenerator.js
+++ b/lib/idGenerator.js
@@ -2,17 +2,38 @@
 
 const crypto = require('crypto')
 
-module.exports = idGenerator(32)
+const randomBytes = crypto.randomBytes
 
-function idGenerator (len) {
-  const fnBody = []
+const cacheSize = 24 << 7
+let pos = 0
+let cache = crypto.randomBytes(cacheSize)
 
-  fnBody.push('const randomInt = crypto.randomInt;')
-  fnBody.push('const base64url = \'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_\'.split(\'\');')
-  fnBody.push('return function () {')
-  const chars = []
-  for (let i = 0; i < ((len) | 0); ++i) chars.push('base64url[randomInt(0, 64)]')
-  fnBody.push('return ' + chars.join('+'))
-  fnBody.push('}')
-  return new Function('crypto', fnBody.join(''))(crypto) // eslint-disable-line no-new-func
+const EQUAL_END_REGEXP = /=/g
+const PLUS_GLOBAL_REGEXP = /\+/g
+const SLASH_GLOBAL_REGEXP = /\//g
+
+module.exports = function (useBase64Url = Buffer.isEncoding('base64url')) {
+  return useBase64Url
+    ? function idGenerator() {
+      if ((pos + 24) > cacheSize) {
+        cache = randomBytes(cacheSize)
+        pos = 0
+      }
+      const buf = Buffer.allocUnsafe(24)
+      cache.copy(buf, 0, pos, (pos += 24))
+      return buf.toString('base64url')
+    }
+    : function idGenerator() {
+      if ((pos + 24) > cacheSize) {
+        cache = randomBytes(cacheSize)
+        pos = 0
+      }
+      const buf = Buffer.allocUnsafe(24)
+      cache.copy(buf, 0, pos, (pos += 24))
+      return buf.toString('base64')
+        .replace(EQUAL_END_REGEXP, '')
+        .replace(PLUS_GLOBAL_REGEXP, '-')
+        .replace(SLASH_GLOBAL_REGEXP, '_')
+    }
 }
+