--- a/lib/idGenerator.js
+++ b/lib/idGenerator.js
@@ -14,7 +14,7 @@ const SLASH_GLOBAL_REGEXP = /\//g
 
 module.exports = function (useBase64Url = Buffer.isEncoding('base64url')) {
   return useBase64Url
-    ? function idGenerator() {
+    ? function idGenerator () {
       if ((pos + 24) > cacheSize) {
         cache = randomBytes(cacheSize)
         pos = 0
@@ -23,7 +23,7 @@ module.exports = function (useBase64Url = Buffer.isEncoding('base64url')) {
       cache.copy(buf, 0, pos, (pos += 24))
       return buf.toString('base64url')
     }
-    : function idGenerator() {
+    : function idGenerator () {
       if ((pos + 24) > cacheSize) {
         cache = randomBytes(cacheSize)
         pos = 0
@@ -36,4 +36,3 @@ module.exports = function (useBase64Url = Buffer.isEncoding('base64url')) {
         .replace(SLASH_GLOBAL_REGEXP, '_')
     }
 }
-