--- a/lib/idGenerator.js
+++ b/lib/idGenerator.js
@@ -1,12 +1,10 @@
 'use strict'
 
-const crypto = require('crypto')
-
-const randomBytes = crypto.randomBytes
+const randomBytes = require('crypto').randomBytes
 
 const cacheSize = 24 << 7
 let pos = 0
-let cache = crypto.randomBytes(cacheSize)
+let cache = randomBytes(cacheSize)
 
 const EQUAL_END_REGEXP = /=/g
 const PLUS_GLOBAL_REGEXP = /\+/g