--- a/types/types.d.ts
+++ b/types/types.d.ts
@@ -120,7 +120,7 @@ declare namespace FastifySessionPlugin {
      */
     rolling?: boolean;
 
-    /** Function used to generate new session IDs. Defaults to uid(24). */
+    /** Function used to generate new session IDs. */
     idGenerator?(request?: Fastify.FastifyRequest): string;
   }