@Uzlopak 2022-08-15
@climba03003 

There are to implications with randomBytes

- randomBytes is not cached but randomInt is cached in nodejs, so perf of randomInt is significantly higher
- base64url is not returning always 32 characters, as trailing `=` (base64-padding)  are removed