@climba03003 2022-08-15
> randomBytes is not cached but randomInt is cached in nodejs, so perf of randomInt is significantly higher

`randomInt` only start out-perform after `5000`, maybe computer dependent.

```js
import crypto from 'crypto'
const randomBytes = crypto.randomBytes

function idGenerator (len) {
  const fnBody = []

  fnBody.push('const randomInt = crypto.randomInt;')
  fnBody.push('const base64url = \'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-_\'.split(\'\');')
  fnBody.push('return function () {')
  const chars = []
  for (let i = 0; i < ((len) | 0); ++i) chars.push('base64url[randomInt(0, 64)]')
  fnBody.push('return ' + chars.join('+'))
  fnBody.push('}')
  return new Function('crypto', fnBody.join(''))(crypto) // eslint-disable-line no-new-func
}

const compute1 = idGenerator(32)
const compute2 = function() {
  return randomBytes(32).toString('base64url')
}

console.time('compute1')
for(let i = 5000; i > 0; i--) {
  compute1()
}
console.timeEnd('compute1')

console.time('compute2')
for(let i = 5000; i > 0; i--) {
  compute2()
}
console.timeEnd('compute2')
```

> base64url is not returning always 32 characters, as trailing = (base64-padding) are removed

The id should be consistence in byte length, not consistence in characters.
This plugin should be compatible with `express-session` store. 
I am not sure if any of the store is storing the `id` as fixed binary length.