@Uzlopak 2022-08-15
I have also a cached randomBytes solution, which is also outperforming non-cached randomBytes.