@Uzlopak 2022-08-15
@climba03003 

```
const crypto = require('crypto')
const randomBytes = crypto.randomBytes

const cacheSize = 24 << 7
let pos = 0
let cache = crypto.randomBytes(cacheSize)

var EQUAL_END_REGEXP = /=/
var PLUS_GLOBAL_REGEXP = /\+/g
var SLASH_GLOBAL_REGEXP = /\//g

const compute1 = Buffer.isEncoding('base64url')
? function () {
  if ((pos + 24) > cacheSize) {
    cache = crypto.randomBytes(cacheSize)
    pos = 0
  }
  const buf = Buffer.allocUnsafe(24)
  cache.copy(buf, 0, pos, (pos += 24))
  return buf.toString('base64url')
}
: function () {
  if ((pos + 24) > cacheSize) {
    cache = crypto.randomBytes(cacheSize)
    pos = 0
  }
  const buf = Buffer.allocUnsafe(24)
  cache.copy(buf, 0, pos, (pos += 24))
  return buf.toString('base64')
    .replace(EQUAL_END_REGEXP, '')
    .replace(PLUS_GLOBAL_REGEXP, '-')
    .replace(SLASH_GLOBAL_REGEXP, '_')
}

const compute2 = Buffer.isEncoding('base64url')
? function() {
  return randomBytes(32).toString('base64url')
}
: function() {
  return randomBytes(32).toString('base64')
  .replace(EQUAL_END_REGEXP, '')
  .replace(PLUS_GLOBAL_REGEXP, '-')
  .replace(SLASH_GLOBAL_REGEXP, '_')
}

console.time('compute1')
for(let i = 50000; i > 0; i--) {
  compute1()
}
console.timeEnd('compute1')

console.time('compute2')
for(let i = 50000; i > 0; i--) {
  compute2()
}
console.timeEnd('compute2')
```

```
uzlopak@Battlestation:~/workspace/session$ nvm use 12
Now using node v12.22.12 (npm v6.14.16)
uzlopak@Battlestation:~/workspace/session$ node test
compute1: 39.415ms
compute2: 80.814ms
uzlopak@Battlestation:~/workspace/session$ nvm use 14
Now using node v14.20.0 (npm v6.14.17)
uzlopak@Battlestation:~/workspace/session$ node test
compute1: 24.481ms
compute2: 69.213ms
uzlopak@Battlestation:~/workspace/session$ nvm use 18
Now using node v18.7.0 (npm v8.15.0)
uzlopak@Battlestation:~/workspace/session$ node test
compute1: 27.079ms
compute2: 114.488ms
```

With 5000 iterations it does not show immediatly, but it is faster in node 18

```
uzlopak@Battlestation:~/workspace/session$ nvm use 12
Now using node v12.22.12 (npm v6.14.16)
uzlopak@Battlestation:~/workspace/session$ node test
compute1: 10.818ms
compute2: 11.453ms
uzlopak@Battlestation:~/workspace/session$ nvm use 14
Now using node v14.20.0 (npm v6.14.17)
uzlopak@Battlestation:~/workspace/session$ node test
compute1: 9.809ms
compute2: 10.247ms
uzlopak@Battlestation:~/workspace/session$ nvm use 16
Now using node v16.16.0 (npm v8.11.0)
uzlopak@Battlestation:~/workspace/session$ node test
compute1: 7.552ms
compute2: 12.115ms
uzlopak@Battlestation:~/workspace/session$ nvm use 18
Now using node v18.7.0 (npm v8.15.0)
uzlopak@Battlestation:~/workspace/session$ node test
compute1: 6.229ms
compute2: 13.205ms
```
And I need to differentiate for base64url because base64url exists since node 14.18.0