@Uzlopak 2022-08-15
If i use cached randombytes i get about 270k-278k ops/s, with randomInt I get constantly above 280k/ops, sometimes spiking to 290k /ops