@Uzlopak 2022-08-15
Ok. got it with cached randomBytes. It is definetly faster than randomInt. Got 277k  ops/s on node 18 vs. randomInt 248k  ops/s vs. uid-safe 212k ops/s